postprocess fragment bug fix
the code checks the difference between 2 long doubles using abs.
however, abs works on the integers, so that also has the side effects of
rounding the diff. in my test stream, the difference is 0.02, and
following the rounding on the call to abs, it becomes 0.
the if block is not entered, and the baseMediaDecodeTime is effectively
ignored, leading to false errors later (e.g. mismatch between sidx
earliest_presentation_time and first frame pts)
the fix is to move to fabsl.